System.Posix

Description

POSIX.1-2008 support

Documentation

module System.Posix.Types

module System.Posix.Signals

module System.Posix.Directory

module System.Posix.Files

module System.Posix.Unistd

module System.Posix.IO

module System.Posix.Env

module System.Posix.Process

module System.Posix.Temp

module System.Posix.Terminal

module System.Posix.Time

module System.Posix.User

module System.Posix.Resource

module System.Posix.Semaphore

module System.Posix.SharedMem

data RTLDFlags

Flags for dlopen.

Constructors

RTLD_LAZY

RTLD_NOW

RTLD_GLOBAL

RTLD_LOCAL

data DL

Flags for dlsym. Notice that Next might not be available on your particular platform! Use haveRtldNext.

If RTLD_DEFAULT is not defined on your platform, packDL Default reduces to nullPtr.

Constructors

Null

Next

DLHandle (Ptr ())

dlopen :: FilePath -> [RTLDFlags] -> IO DL

dlsym :: DL -> String -> IO (FunPtr a)

dlsym returns the address binding of the symbol described in symbol, as it occurs in the shared object identified by source.

dlerror :: IO String

dlclose :: DL -> IO ()

withDL :: FilePath -> [RTLDFlags] -> (DL -> IO a) -> IO a

withDL_ :: FilePath -> [RTLDFlags] -> (DL -> IO a) -> IO ()

undl :: DL -> Ptr ()

undl obtains the raw handle. You mustn't do something like withDL mod flags $ liftM undl >>= p -> use p

c_dlopen :: CString -> CInt -> IO (Ptr ())

c_dlsym :: Ptr () -> CString -> IO (FunPtr a)

c_dlerror :: IO CString

c_dlclose :: Ptr () -> IO CInt

haveRtldNext :: Bool

On some hosts (e.g. SuSe and Ubuntu Linux) RTLD_NEXT (and RTLD_DEFAULT) are not visible without setting the macro _GNU_SOURCE. Since we don't want to define this macro, you can use the function haveRtldNext to check whether the flag Next is available. Ideally, this will be optimized by the compiler so that it should be as efficient as an #ifdef.

If you fail to test the flag and use it although it is undefined, packDL will throw an error.

haveRtldLocal :: Bool

Deprecated: defaults to True

packRTLDFlags :: [RTLDFlags] -> CInt

packDL :: DL -> Ptr ()