System.Posix.ByteString

Description

POSIX.1-2008 support with ByteString file paths and environment strings.

This module exports exactly the same API as System.Posix, except that all file paths and environment strings are represented by ByteString instead of String. The System.Posix API implicitly translates all file paths and environment strings using the locale encoding, whereas this version of the API does no encoding or decoding and works directly in terms of raw bytes.

Note that if you do need to interpret file paths or environment strings as text, then some Unicode encoding or decoding should be applied first.

Documentation

type RawFilePath = ByteString

A literal POSIX file path

module System.Posix.Types

module System.Posix.Signals

module System.Posix.Directory.ByteString

module System.Posix.Files.ByteString

module System.Posix.Unistd

module System.Posix.IO.ByteString

module System.Posix.Env.ByteString

module System.Posix.Process.ByteString

module System.Posix.Temp.ByteString

module System.Posix.Terminal.ByteString

module System.Posix.Time

module System.Posix.User

module System.Posix.Resource

module System.Posix.Semaphore

module System.Posix.SharedMem

data RTLDFlags

Flags for dlopen.

Constructors

RTLD_LAZY

RTLD_NOW

RTLD_GLOBAL

RTLD_LOCAL

data DL

Flags for dlsym. Notice that Next might not be available on your particular platform! Use haveRtldNext.

If RTLD_DEFAULT is not defined on your platform, packDL Default reduces to nullPtr.

Constructors

Null

Next

DLHandle (Ptr ())

dlopen :: RawFilePath -> [RTLDFlags] -> IO DL

dlsym :: DL -> String -> IO (FunPtr a)

dlsym returns the address binding of the symbol described in symbol, as it occurs in the shared object identified by source.

dlerror :: IO String

dlclose :: DL -> IO ()

withDL :: RawFilePath -> [RTLDFlags] -> (DL -> IO a) -> IO a

withDL_ :: RawFilePath -> [RTLDFlags] -> (DL -> IO a) -> IO ()

undl :: DL -> Ptr ()

undl obtains the raw handle. You mustn't do something like withDL mod flags $ liftM undl >>= p -> use p

c_dlopen :: CString -> CInt -> IO (Ptr ())

c_dlsym :: Ptr () -> CString -> IO (FunPtr a)

c_dlerror :: IO CString

c_dlclose :: Ptr () -> IO CInt

haveRtldNext :: Bool

On some hosts (e.g. SuSe and Ubuntu Linux) RTLD_NEXT (and RTLD_DEFAULT) are not visible without setting the macro _GNU_SOURCE. Since we don't want to define this macro, you can use the function haveRtldNext to check whether the flag Next is available. Ideally, this will be optimized by the compiler so that it should be as efficient as an #ifdef.

If you fail to test the flag and use it although it is undefined, packDL will throw an error.

haveRtldLocal :: Bool

Deprecated: defaults to True

packRTLDFlags :: [RTLDFlags] -> CInt

packDL :: DL -> Ptr ()