System.Posix.Directory

Description

String-based POSIX directory support

Creating and removing directories

createDirectory :: FilePath -> FileMode -> IO ()

createDirectory dir mode calls mkdir to create a new directory, dir, with permissions based on mode.

removeDirectory :: FilePath -> IO ()

Reading directories

data DirStream

openDirStream :: FilePath -> IO DirStream

openDirStream dir calls opendir to obtain a directory stream for dir.

readDirStream :: DirStream -> IO FilePath

readDirStream dp calls readdir to obtain the next directory entry (struct dirent) for the open directory stream dp, and returns the d_name member of that structure.

Note that this function returns an empty filepath if the end of the directory stream is reached. For a safer alternative use readDirStreamMaybe.

readDirStreamMaybe :: DirStream -> IO (Maybe FilePath)

readDirStreamMaybe dp calls readdir to obtain the next directory entry (struct dirent) for the open directory stream dp. It returns the d_name member of that structure wrapped in a Just d_name if an entry was read and Nothing if the end of the directory stream was reached.

rewindDirStream :: DirStream -> IO ()

rewindDirStream dp calls rewinddir to reposition the directory stream dp at the beginning of the directory.

closeDirStream :: DirStream -> IO ()

closeDirStream dp calls closedir to close the directory stream dp.

data DirStreamOffset

tellDirStream :: DirStream -> IO DirStreamOffset

seekDirStream :: DirStream -> DirStreamOffset -> IO ()

The working dirctory

getWorkingDirectory :: IO FilePath

getWorkingDirectory calls getcwd to obtain the name of the current working directory.

changeWorkingDirectory :: FilePath -> IO ()

changeWorkingDirectory dir calls chdir to change the current working directory to dir.

changeWorkingDirectoryFd :: Fd -> IO ()