System.Posix.Env.ByteString

Description

POSIX environment support

Environment Variables

getEnv

Arguments

:: ByteString
variable name

-> IO (Maybe ByteString)
variable value

getEnv looks up a variable in the environment.

getEnvDefault

Arguments

:: ByteString
variable name

-> ByteString
fallback value

-> IO ByteString
variable value or fallback value

getEnvDefault is a wrapper around getEnv where the programmer can specify a fallback as the second argument, which will be used if the variable is not found in the environment.

getEnvironmentPrim :: IO [ByteString]

getEnvironment

Arguments

:: IO [(ByteString, ByteString)]
[(key,value)]

getEnvironment retrieves the entire environment as a list of (key,value) pairs.

setEnvironment

Arguments

:: [(ByteString, ByteString)]
[(key,value)]

-> IO ()

setEnvironment resets the entire environment to the given list of (key,value) pairs.

Since: unix-2.8.0.0

putEnv

Arguments

:: ByteString
"key=value"

-> IO ()

putEnv function takes an argument of the form name=value and is equivalent to setEnv(key,value,True{-overwrite-}).

setEnv

Arguments

:: ByteString
variable name

-> ByteString
variable value

-> Bool
overwrite

-> IO ()

The setEnv function inserts or resets the environment variable name in the current environment list. If the variable name does not exist in the list, it is inserted with the given value. If the variable does exist, the argument overwrite is tested; if overwrite is False, the variable is not reset, otherwise it is reset to the given value.

unsetEnv

Arguments

:: ByteString
variable name

-> IO ()

The unsetEnv function deletes all instances of the variable name from the environment.

clearEnv :: IO ()

The clearEnv function clears the environment of all name-value pairs.

Program arguments

getArgs :: IO [ByteString]

Computation getArgs returns a list of the program's command line arguments (not including the program name), as ByteStrings.

Unlike getArgs, this function does no Unicode decoding of the arguments; you get the exact bytes that were passed to the program by the OS. To interpret the arguments as text, some Unicode decoding should be applied.