System.Posix.PosixPath.FilePath

Description

Internal stuff: support for ByteString FilePaths

Documentation

withFilePath :: PosixPath -> (CString -> IO a) -> IO a

peekFilePath :: CString -> IO PosixPath

peekFilePathLen :: CStringLen -> IO PosixPath

throwErrnoPathIfMinus1Retry :: (Eq a, Num a) => String -> PosixPath -> IO a -> IO a

throwErrnoPathIfMinus1Retry_ :: (Eq a, Num a) => String -> PosixPath -> IO a -> IO ()

throwErrnoPathIfNullRetry :: String -> PosixPath -> IO (Ptr a) -> IO (Ptr a)

throwErrnoPathIfRetry :: (a -> Bool) -> String -> PosixPath -> IO a -> IO a

throwErrnoPath :: String -> PosixPath -> IO a

as throwErrno, but exceptions include the given path when appropriate.

throwErrnoPathIf :: (a -> Bool) -> String -> PosixPath -> IO a -> IO a

as throwErrnoIf, but exceptions include the given path when appropriate.

throwErrnoPathIf_ :: (a -> Bool) -> String -> PosixPath -> IO a -> IO ()

as throwErrnoIf_, but exceptions include the given path when appropriate.

throwErrnoPathIfNull :: String -> PosixPath -> IO (Ptr a) -> IO (Ptr a)

as throwErrnoIfNull, but exceptions include the given path when appropriate.

throwErrnoPathIfMinus1 :: (Eq a, Num a) => String -> PosixPath -> IO a -> IO a

as throwErrnoIfMinus1, but exceptions include the given path when appropriate.

throwErrnoPathIfMinus1_ :: (Eq a, Num a) => String -> PosixPath -> IO a -> IO ()

as throwErrnoIfMinus1_, but exceptions include the given path when appropriate.

throwErrnoTwoPathsIfMinus1_ :: (Eq a, Num a) => String -> PosixPath -> PosixPath -> IO a -> IO ()

as throwErrnoTwoPathsIfMinus1_, but exceptions include two paths when appropriate.