System.Posix.PosixString

Description

POSIX.1-2008 support with ByteString file paths and environment strings.

This module exports exactly the same API as System.Posix, except that all file paths and environment strings are represented by ByteString instead of String. The System.Posix API implicitly translates all file paths and environment strings using the locale encoding, whereas this version of the API does no encoding or decoding and works directly in terms of raw bytes.

Note that if you do need to interpret file paths or environment strings as text, then some Unicode encoding or decoding should be applied first.

Documentation

data PosixString

Commonly used Posix string as uninterpreted char[] array.

type PosixPath = PosixString

Filepaths are char[] data on unix as passed to syscalls.

module System.Posix.Types

module System.Posix.Signals

module System.Posix.Directory.PosixPath

module System.Posix.Files.PosixString

module System.Posix.Unistd

module System.Posix.IO.PosixString

module System.Posix.Env.PosixString

module System.Posix.Process.PosixString

module System.Posix.Temp.PosixString

module System.Posix.Time

module System.Posix.User

module System.Posix.Resource

module System.Posix.Semaphore

module System.Posix.SharedMem