System.Posix.Process.Internals

Documentation

pPrPr_disableITimers :: IO ()

c_execvpe :: CString -> Ptr CString -> Ptr CString -> IO CInt

decipherWaitStatus :: CInt -> IO ProcessStatus

data ProcessStatus

The exit status of a process

Constructors

Exited ExitCode
the process exited by calling exit() or returning from main

Terminated Signal Bool
the process was terminated by a signal, the Bool is True if a core dump was produced

Since: unix-2.7.0.0

Stopped Signal
the process was stopped by a signal