System.Posix.SharedMem

Description

POSIX shared memory support.

Documentation

data ShmOpenFlags

Constructors

ShmOpenFlags

Fields

shmReadWrite :: Bool

If true, open the shm object read-write rather than read-only.

shmCreate :: Bool

If true, create the shm object if it does not exist.

shmExclusive :: Bool

If true, throw an exception if the shm object already exists.

shmTrunc :: Bool

If true, wipe the contents of the shm object after opening it.

shmOpen :: String -> ShmOpenFlags -> FileMode -> IO Fd

Open a shared memory object with the given name, flags, and mode.

shmUnlink :: String -> IO ()

Delete the shared memory object with the given name.