System.Posix.Signals

Description

POSIX signal support

The Signal type

type Signal = CInt

Specific signals

nullSignal :: Signal

internalAbort :: Signal

Alias for sigABRT.

sigABRT :: CInt

Process abort signal.

realTimeAlarm :: Signal

Alias for sigALRM.

sigALRM :: CInt

Alarm clock.

busError :: Signal

Alias for sigBUS.

sigBUS :: CInt

Access to an undefined portion of a memory object.

processStatusChanged :: Signal

Alias for sigCHLD.

sigCHLD :: CInt

Child process terminated, stopped, or continued.

continueProcess :: Signal

Alias for sigCONT.

sigCONT :: CInt

Continue executing, if stopped.

floatingPointException :: Signal

Alias for sigFPE.

sigFPE :: CInt

Erroneous arithmetic operation.

lostConnection :: Signal

Alias for sigHUP.

sigHUP :: CInt

Hangup.

illegalInstruction :: Signal

Alias for sigILL.

sigILL :: CInt

Illegal instruction.

keyboardSignal :: Signal

Alias for sigINT.

sigINT :: CInt

Terminal interrupt signal.

killProcess :: Signal

Alias for sigKILL.

sigKILL :: CInt

Kill (cannot be caught or ignored).

openEndedPipe :: Signal

Alias for sigPIPE.

sigPIPE :: CInt

Write on a pipe with no one to read it.

keyboardTermination :: Signal

Alias for sigQUIT.

sigQUIT :: CInt

Terminal quit signal.

segmentationViolation :: Signal

Alias for sigSEGV.

sigSEGV :: CInt

Invalid memory reference.

softwareStop :: Signal

Alias for sigSTOP.

sigSTOP :: CInt

Stop executing (cannot be caught or ignored).

softwareTermination :: Signal

Alias for sigTERM.

sigTERM :: CInt

Termination signal.

keyboardStop :: Signal

Alias for sigTSTP.

sigTSTP :: CInt

Terminal stop signal.

backgroundRead :: Signal

Alias for sigTTIN.

sigTTIN :: CInt

Background process attempting read.

backgroundWrite :: Signal

Alias for sigTTOU.

sigTTOU :: CInt

Background process attempting write.

userDefinedSignal1 :: Signal

Alias for sigUSR1.

sigUSR1 :: CInt

User-defined signal 1.

userDefinedSignal2 :: Signal

Alias for sigUSR2.

sigUSR2 :: CInt

User-defined signal 2.

pollableEvent :: Signal

Alias for sigPOLL.

sigPOLL :: CInt

Pollable event.

profilingTimerExpired :: Signal

Alias for sigPROF.

sigPROF :: CInt

Profiling timer expired.

badSystemCall :: Signal

Alias for sigSYS.

sigSYS :: CInt

Bad system call.

breakpointTrap :: Signal

Alias for sigTRAP.

sigTRAP :: CInt

Trace/breakpoint trap.

urgentDataAvailable :: Signal

Alias for sigURG.

sigURG :: CInt

High bandwidth data is available at a socket.

virtualTimerExpired :: Signal

Alias for sigVTALRM.

sigVTALRM :: CInt

Virtual timer expired.

cpuTimeLimitExceeded :: Signal

Alias for sigXCPU.

sigXCPU :: CInt

CPU time limit exceeded.

fileSizeLimitExceeded :: Signal

Alias for sigXFSZ.

sigXFSZ :: CInt

File size limit exceeded.

Sending signals

raiseSignal :: Signal -> IO ()

raiseSignal int calls kill to signal the current process with interrupt signal int.

signalProcess :: Signal -> ProcessID -> IO ()

signalProcess int pid calls kill to signal process pid with interrupt signal int.

signalProcessGroup :: Signal -> ProcessGroupID -> IO ()

signalProcessGroup int pgid calls kill to signal all processes in group pgid with interrupt signal int.

Handling signals

data Handler

The actions to perform when a signal is received.

Constructors

Default
Sets the disposition of the signal to SIG_DFL, which means we want the default action associated with the signal. For example, the default action for SIGTERM (and various other signals) is to terminate the process.

Ignore
Set the disposition of the signal to SIG_IGN, which means we want to ignore the signal. Ignored signals will not be delivered to the process, and if also blocked will not be added to the pending set for later delivery (if/when unblocked). Some signals (e.g. SIGSTOP and SIGKILL) cannot be caught or ignored. not yet: | Hold

Catch (IO ())
signal handler is not reset

CatchOnce (IO ())
signal handler is automatically reset (via SA_RESETHAND)

CatchInfo (SignalInfo -> IO ())
Since: unix-2.7.0.0

CatchInfoOnce (SignalInfo -> IO ())
Since: unix-2.7.0.0

data SignalInfo

Information about a received signal (derived from siginfo_t).

Since: unix-2.7.0.0

Constructors

SignalInfo

Fields

siginfoSignal :: Signal

siginfoError :: Errno

siginfoSpecific :: SignalSpecificInfo

data SignalSpecificInfo

Information specific to a particular type of signal (derived from siginfo_t).

Since: unix-2.7.0.0

Constructors

NoSignalSpecificInfo

SigChldInfo

Fields

siginfoPid :: ProcessID

siginfoUid :: UserID

siginfoStatus :: ProcessStatus

installHandler

Arguments

:: Signal

-> Handler

-> Maybe SignalSet
other signals to block

-> IO Handler
old handler

installHandler int handler iset calls sigaction to install an interrupt handler for signal int. If handler is Default, SIG_DFL is installed; if handler is Ignore, SIG_IGN is installed; if handler is Catch action, a handler is installed which will invoke action in a new thread when (or shortly after) the signal is received. If iset is Just s, then the sa_mask of the sigaction structure is set to s; otherwise it is cleared. The previously installed signal handler for int is returned

Signal sets

data SignalSet

emptySignalSet :: SignalSet

fullSignalSet :: SignalSet

reservedSignals :: SignalSet

A set of signals reserved for use by the implementation. In GHC, this will normally include either sigVTALRM or sigALRM.

addSignal :: Signal -> SignalSet -> SignalSet infixr 9

deleteSignal :: Signal -> SignalSet -> SignalSet infixr 9

inSignalSet :: Signal -> SignalSet -> Bool

The process signal mask

getSignalMask :: IO SignalSet

getSignalMask calls sigprocmask to determine the set of interrupts which are currently being blocked.

setSignalMask :: SignalSet -> IO ()

setSignalMask mask calls sigprocmask with SIG_SETMASK to block all interrupts in mask.

blockSignals :: SignalSet -> IO ()

blockSignals mask calls sigprocmask with SIG_BLOCK to add all interrupts in mask to the set of blocked interrupts.

unblockSignals :: SignalSet -> IO ()

unblockSignals mask calls sigprocmask with SIG_UNBLOCK to remove all interrupts in mask from the set of blocked interrupts.

The alarm timer

scheduleAlarm :: Int -> IO Int

scheduleAlarm i calls alarm to schedule a real time alarm at least i seconds in the future.

Waiting for signals

getPendingSignals :: IO SignalSet

getPendingSignals calls sigpending to obtain the set of interrupts which have been received but are currently blocked.

awaitSignal :: Maybe SignalSet -> IO ()

awaitSignal iset suspends execution until an interrupt is received. If iset is Just s, awaitSignal calls sigsuspend, installing s as the new signal mask before suspending execution; otherwise, it calls sigsuspend with current signal mask. Note that RTS scheduler signal (either virtualTimerExpired or realTimeAlarm) could cause premature termination of this call. It might be necessary to block that signal before invocation of awaitSignal with blockSignals reservedSignals.

awaitSignal returns when signal was received and processed by a signal handler, or if the signal could not be caught. If you have installed any signal handlers with installHandler, it may be wise to call yield directly after awaitSignal to ensure that the signal handler runs as promptly as possible.

The NOCLDSTOP flag

setStoppedChildFlag :: Bool -> IO Bool

Tells the system whether or not to set the SA_NOCLDSTOP flag when installing new signal handlers.

queryStoppedChildFlag :: IO Bool

Queries the current state of the stopped child flag.