System.Posix.Temp

Description

POSIX environment support

Documentation

mkstemp :: String -> IO (FilePath, Handle)

mkstemp - make a unique filename and open it for reading/writing (only safe on GHC & Hugs). The returned FilePath is the (possibly relative) path of the created file, which is padded with 6 random characters.