System.Posix.Temp.PosixString

Description

POSIX temporary file and directory creation functions.

Documentation

mkstemp :: PosixString -> IO (PosixPath, Handle)

Make a unique filename and open it for reading/writing. The returned PosixPath is the (possibly relative) path of the created file, which is padded with 6 random characters. The argument is the desired prefix of the filepath of the temporary file to be created.

If you aren't using GHC or Hugs then this function simply wraps mktemp and so shouldn't be considered safe.

mkstemps :: PosixString -> PosixString -> IO (PosixPath, Handle)

mkstemps - make a unique filename with a given prefix and suffix and open it for reading/writing (only safe on GHC & Hugs). The returned PosixPath is the (possibly relative) path of the created file, which contains 6 random characters in between the prefix and suffix.

mkdtemp :: PosixString -> IO PosixPath

Make a unique directory. The returned PosixPath is the path of the created directory, which is padded with 6 random characters. The argument is the desired prefix of the filepath of the temporary directory to be created.

If you aren't using GHC or Hugs then this function simply wraps mktemp and so shouldn't be considered safe.