System.Posix.Unistd

Description

POSIX miscellaneous stuff, mostly from unistd.h

System environment

data SystemID

Constructors

SystemID

Fields

systemName :: String

nodeName :: String

release :: String

version :: String

machine :: String

getSystemID :: IO SystemID

data SysVar

Constructors

ArgumentLimit

ChildLimit

ClockTick

GroupLimit

OpenFileLimit

PosixVersion

HasSavedIDs

HasJobControl

getSysVar :: SysVar -> IO Integer

Sleeping

sleep :: Int -> IO Int

Sleep for the specified duration (in seconds). Returns the time remaining (if the sleep was interrupted by a signal, for example).

GHC Note: the comment for usleep also applies here.

usleep :: Int -> IO ()

Sleep for the specified duration (in microseconds).

GHC Note: threadDelay is a better choice. Without the -threaded option, usleep will block all other user threads. Even with the -threaded option, usleep requires a full OS thread to itself. threadDelay has neither of these shortcomings.

nanosleep :: Integer -> IO ()

Sleep for the specified duration (in nanoseconds)