System.Posix.User.ByteString

Description

POSIX user/group support

User environment

Querying the user environment

getRealUserID :: IO UserID

getRealUserID calls getuid to obtain the real UserID associated with the current process.

getRealGroupID :: IO GroupID

getRealGroupID calls getgid to obtain the real GroupID associated with the current process.

getEffectiveUserID :: IO UserID

getEffectiveUserID calls geteuid to obtain the effective UserID associated with the current process.

getEffectiveGroupID :: IO GroupID

getEffectiveGroupID calls getegid to obtain the effective GroupID associated with the current process.

getGroups :: IO [GroupID]

getGroups calls getgroups to obtain the list of supplementary GroupIDs associated with the current process.

getLoginName :: IO ByteString

getLoginName calls getlogin to obtain the login name associated with the current process.

getEffectiveUserName :: IO ByteString

getEffectiveUserName gets the name associated with the effective UserID of the process.

The group database

data GroupEntry

Constructors

GroupEntry

Fields

groupName :: ByteString

The name of this group (gr_name)

groupPassword :: ByteString

The password for this group (gr_passwd)

groupID :: GroupID

The unique numeric ID for this group (gr_gid)

groupMembers :: [ByteString]

A list of zero or more usernames that are members (gr_mem)

getGroupEntryForID :: GroupID -> IO GroupEntry

getGroupEntryForID gid calls getgrgid_r to obtain the GroupEntry information associated with GroupID gid. This operation may fail with isDoesNotExistError if no such group exists.

getGroupEntryForName :: ByteString -> IO GroupEntry

getGroupEntryForName name calls getgrnam_r to obtain the GroupEntry information associated with the group called name. This operation may fail with isDoesNotExistError if no such group exists.

getAllGroupEntries :: IO [GroupEntry]

getAllGroupEntries returns all group entries on the system by repeatedly calling getgrent

The user database

data UserEntry

Constructors

UserEntry

Fields

userName :: ByteString

Textual name of this user (pw_name)

userPassword :: ByteString

Password -- may be empty or fake if shadow is in use (pw_passwd)

userID :: UserID

Numeric ID for this user (pw_uid)

userGroupID :: GroupID

Primary group ID (pw_gid)

userGecos :: ByteString

Usually the real name for the user (pw_gecos)

homeDirectory :: ByteString

Home directory (pw_dir)

userShell :: ByteString

Default shell (pw_shell)

getUserEntryForID :: UserID -> IO UserEntry

getUserEntryForID uid calls getpwuid_r to obtain the UserEntry information associated with UserID uid. This operation may fail with isDoesNotExistError if no such user exists.

getUserEntryForName :: ByteString -> IO UserEntry

getUserEntryForName name calls getpwnam_r to obtain the UserEntry information associated with the user login name. This operation may fail with isDoesNotExistError if no such user exists.

getAllUserEntries :: IO [UserEntry]

getAllUserEntries returns all user entries on the system by repeatedly calling getpwent

Modifying the user environment

setUserID :: UserID -> IO ()

setUserID uid calls setuid to set the real, effective, and saved set-user-id associated with the current process to uid.

setGroupID :: GroupID -> IO ()

setGroupID gid calls setgid to set the real, effective, and saved set-group-id associated with the current process to gid.

setEffectiveUserID :: UserID -> IO ()

setEffectiveUserID uid calls seteuid to set the effective user-id associated with the current process to uid. This does not update the real user-id or set-user-id.

setEffectiveGroupID :: GroupID -> IO ()

setEffectiveGroupID uid calls setegid to set the effective group-id associated with the current process to gid. This does not update the real group-id or set-group-id.

setGroups :: [GroupID] -> IO ()

setGroups calls setgroups to set the list of supplementary GroupIDs associated with the current process.